правки для следующего модуля
