+++ js/check-line.js
@@ -1,0 +1,10 @@
+function checkLineWidth(string, number) {
+  if (string <= number) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+export {checkLineWidth};
+
